requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# Cryptography (RSA keys generation and JWT signing)
cryptography>=41.0.0

# Email validation (for user registration)
email-validator==2.1.0

# HTTP client (for testing)
httpx==0.25.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
